requirements.txt
accelerate
aiofiles
aistudio-sdk
alabaster
alembic
altair
altgraph
amqp
aniso8601
annotated-types
antlr4-python3-runtime
anyio
APScheduler
arabic-reshaper
asn1crypto
astor
attrs
babel
backoff
bce-python-sdk
bcrypt
beautifulsoup4
bidict
billiard
black
blinker
boto3
botocore
brotli
cachetools
camelot-py
celery
certifi
cffi
chardet
charset-normalizer
click
click-didyoumean
click-plugins
click-repl
colorama
coloredlogs
colorlog
comtypes
configparser
contourpy
coverage
cryptography
cssselect2
cycler
dataclasses-json
decorator
defusedxml
Deprecated
distro
dnspython
docutils
dropbox
easyocr
effdet
email-validator
emoji
et_xmlfile
excalibur-py
faiss-cpu
fastapi
filelock
filetype
flake8
Flask
flask-cors
Flask-Login
Flask-RESTful
Flask-SocketIO
Flask-SQLAlchemy
Flask-WTF
flatbuffers
fonttools
fpdf2
freetype-py
fsspec
ftfy
future
gitdb
GitPython
google-api-core
google-api-python-client
google-auth
google-auth-httplib2
google-auth-oauthlib
google-cloud-vision
googleapis-common-protos
greenlet
grpcio
grpcio-status
gunicorn
h11
html5lib
httpcore
httplib2
httptools
httpx
huggingface-hub
humanfriendly
idna
imageio
imagesize
iniconfig
iopath
itsdangerous
Jinja2
jiter
jmespath
joblib
jsonschema
jsonschema-specifications
kiwisolver
kombu
langdetect
layoutparser
lazy_loader
Levenshtein
loguru
lxml
Mako
Markdown
markdown-it-py
MarkupSafe
marshmallow
matplotlib
mccabe
mdurl
ml_dtypes
modelscope
mpmath
msal
msoffcrypto-tool
mypy_extensions
narwhals
networkx
ninja
nltk
numpy
oauthlib
Office365-REST-Python-Client
olefile
omegaconf
onnxruntime
open_clip_torch
openai
opencv-contrib-python
opencv-python
opencv-python-headless
openpyxl
opt-einsum
oscrypto
packaging
paddleocr
paddlepaddle
paddlex
pandas
pathspec
pdf2image
pdfminer.six
pdfplumber
pefile
pi_heif
pikepdf
pillow
platformdirs
plotly
pluggy
ply
portalocker
prettytable
prometheus-client
prompt_toolkit
proto-plus
protobuf
psutil
psycopg2-binary
pwinput
py-cpuinfo
pyarrow
pyasn1
pyasn1_modules
pycairo
pyclipper
pycocotools
pycodestyle
pycparser
pycryptodome
pydantic-settings
pydantic_core
pydeck
pydyf
pyflakes
Pygments
pyHanko
pyhanko-certvalidator
pyinstaller
pyinstaller-hooks-contrib
PyJWT
PyMuPDF
PyMuPDFb
pyodbc
pyotp
pypandoc
pyparsing
pypdf
pypdfium2
pyphen
pyreadline3
pytesseract
pytest
pytest-cov
python-bidi
python-dateutil
python-decouple
python-docx
python-dotenv
python-engineio
python-iso639
python-Levenshtein
python-magic
python-multipart
python-oxmsg
python-pptx
python-socketio
pytokens
pytwain
pytz
pywin32
pywin32-ctypes
PyYAML
RapidFuzz
referencing
regex
reportlab
requests
requests-oauthlib
requests-toolbelt
rich
rlPyCairo
roman-numerals-py
rpds-py
rsa
ruamel.yaml
ruamel.yaml.clib
s3transfer
safetensors
schedule
scikit-image
scikit-learn
scipy
sentence-transformers
sentencepiece
setuptools
shapely
simple-websocket
six
smmap
sniffio
snowballstemmer
soupsieve
Sphinx
sphinxcontrib-applehelp
sphinxcontrib-devhelp
sphinxcontrib-htmlhelp
sphinxcontrib-jsmath
sphinxcontrib-qthelp
sphinxcontrib-serializinghtml
SQLAlchemy
starlette
stone
streamlit
streamlit-aggrid
streamlit-option-menu
svglib
sympy
tabula-py
tabulate
tenacity
threadpoolctl
tifffile
timm
tinycss2
tinyhtml5
tokenizers
toml
torch
torchaudio
torchvision
tornado
tqdm
transformers
typing-inspect
typing-inspection
typing_extensions
tzdata
tzlocal
ujson
uritemplate
uritools
urllib3
uvicorn
vine
watchdog
watchfiles
wcwidth
weasyprint
webencodings
websockets
Werkzeug
wheel
win32_setctime
wrapt
wsproto
WTForms
xhtml2pdf
xlrd
xlsxwriter
zopfli
Flask-Limiter
pydantic
reportlab
